#rb none

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: chris.gagnon
#ROBOMERGE-SOURCE: CL 11224363 in //UE4/Main/...
#ROBOMERGE-BOT: NETWORKING (Main -> Dev-Networking) (v654-11333218)

[CL 11472336 by chris gagnon in Dev-Networking branch]

+++ Engine/Plugins/Messaging/UdpMessaging/Source/UdpMessaging/Private/Transport/UdpMessageSegmenter.cpp
@@ -44,7 +44,7 @@ bool FUdpMessageSegmenter::GetNextPendingSegment(TArray<uint8>& OutData, uint32&
 	{
 		OutSegment = It.GetIndex();
 
-		uint64 SegmentOffset = OutSegment * SegmentSize;
+		uint64 SegmentOffset = static_cast<uint64>(OutSegment) * SegmentSize;
 		uint64 ActualSegmentSize = MessageReader->TotalSize() - SegmentOffset;
 
 		if (ActualSegmentSize > SegmentSize)
@@ -77,7 +77,7 @@ bool FUdpMessageSegmenter::GetPendingSegment(uint32 InSegment, TArray<uint8>& Ou
 	// Max segment number for protocol 12 is INT32_MAX, if increased, this will need changing
 	if (InSegment < (uint32)PendingSegments.Num() && PendingSegments[InSegment])
 	{
-		uint64 SegmentOffset = InSegment * SegmentSize;
+		uint64 SegmentOffset = static_cast<uint64>(InSegment) * SegmentSize;
 		uint64 ActualSegmentSize = MessageReader->TotalSize() - SegmentOffset;
 
 		if (ActualSegmentSize > SegmentSize)
